merged

+++ Dimitri/Game/Game/src/Models/Buttons/StartButton.cpp
@@ -3,7 +3,7 @@
 
 namespace Game {
 	namespace Models {
-		Buttons::StartButton::StartButton(int x, int y, int z, int height, int width, Enums::DirectionEnum state, int scene_height, std::shared_ptr<Managers::LevelManager> level_manager, Graphics::Models::Center center) :
+		Buttons::StartButton::StartButton(float x, float y, float z, float height, float width, Enums::DirectionEnum state, float scene_height, std::shared_ptr<Managers::LevelManager> level_manager, Graphics::Models::Center center) :
 			Button(x, y, z, height, width, state, scene_height, center, "StartButton")
 		{
 			_level_manager = level_manager;
